Revert "Memoize entity encoding"
This reverts commit f5bae9ea9c3e30793e83a3f50a0172bae614ae2c.

Because we cannot forsee the length and amount of strings we need to
be more careful about keeping them around as memory consumption
can grow unbounded. In the future we should check if we can use
a limited cache, only cache attributes or something else. WeakMaps
unfortunately won't help as they cannot be used with primitives as
cache key.